Action Bar support added.
Also trying to use ShareActionProvider, but it has one very bad feature - it inserts a new icon into the action bar when a share item is clicked. Tried multiple ways to suppress it, nothing worked. So this is just a temporary checkin. Will remove it next checkin, and just handle the share event as we did before this checkin.