Need to avoid allowing Travel Blog to start in another task.
Because Travel Blog always opens the last opened trip, it will cause incorrect trips to be loaded if any travel blog activity starts in some other task.
Example scenario:
Travel Blog is editing a trip A. Then starts a New Post.
Then hits home screen, and starts some other activity.
Then later, use the launch icon to start Travel Blog again.
Android will reuse the old task, and bring the old task to front - this is the standard behavior.
But if instead of restart Travel Blog from launcher, say user starts a File Manager app, and clicks on some .kml file, and then  Sends it to Travel Blog.
This starts the Import activity of Travel Blog, but this activity is part of the file manager task. If the user now clicks on Save import, then we want to start Travel Blog with the newly imported file.
But this is a new task, new process. There is an old travel blog process in the back stack.
Since the model for Travel Blog is to save the last used file, having two travel blogs with two different files in the back stack cannot be supported.
So, when we launch Travel Blog from the Import activity, we use the flags Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_NEW_TASK which reuses the old travel blog process (if present), and thus avoids the problem with having two processes.