Continued improvements related to multiple tasks.
Now that the Import Trip capability can start Travel Blog activity screens in non-Travel Blog tasks, need to be careful about how the last opened file preferences is saved, as well as need to make sure that each activity keeps track of which blog it is operating on.

There are some holes - since it is now possible to have a back stack like this of Travel Blog activities:

TravelLocBlogMain -> EditBlogElement ... -> TravelLogBlogMain -> EditBlogElement
Now if the second Main activity edits or deletes the blog being editing in the first Main and Edit activities, when user hits the back button, they will not be really working on valid data. No easy fix possible - this is not an issue for the most normal use cases where users complete a edit task (Save or Cancel), etc.
If user leaves Edit screen up and clicks on Home button, then there is a chance of some confusion.
Only way to really fix is this to use the new Android model of always saving the data as soon as user transitions out of the Edit activity (onPause). But that also requires an Undo capability. And Android does not have a text-edit undo capability yet, so always saving data is not really a good user experience.

The current checkin seems good enough for now, may try some more minor enhancements in this area.